sharedPtr fixed, btree fully implemented w/o tests yet

+++ inc/utilInc/ArraySequence.hpp
@@ -9,7 +9,7 @@ class ArraySequence : public Sequence<T>
 {
 public:
     ArraySequence();
-    ArraySequence( const int capacity );
+    ArraySequence( const size_t capacity );
 
     ArraySequence( const DynamicArray<T>& src );
     ArraySequence( const ArraySequence<T>& src );
@@ -29,28 +29,28 @@ public:
 public:
     void append( const T& value ) override;
     void prepend( const T& value ) override;
-    void insertAt( const T& value, const int pos ) override;
-    void removeAt( const int pos ) override;
-    void setAt( const T& value, const int pos ) override;
-    void swap( const int pos1, const int pos2 ) override;
-    ArraySequence<T> subArray( const int startIndex, const int endIndex ) const;
-    Sequence<T>* getSubSequence( const int startIndex, const int endIndex ) const override;
+    void insertAt( const T& value, const size_t pos ) override;
+    void removeAt( const size_t pos ) override;
+    void setAt( const T& value, const size_t pos ) override;
+    void swap( const size_t pos1, const size_t pos2 ) override;
+    ArraySequence<T> subArray( const size_t startIndex, const size_t endIndex ) const;
+    Sequence<T>* getSubSequence( const size_t startIndex, const size_t endIndex ) const override;
     Sequence<T>* concat( const Sequence<T>& other ) override;
     void map( const std::function<T(T)>& func );
     void where( const std::function<bool(T)>& func );
 public:
-    T& operator[]( const int pos ) override;
-    const T& operator[]( const int pos ) const override;
+    T& operator[]( const size_t pos ) override;
+    const T& operator[]( const size_t pos ) const override;
 public:
     bool isEmpty() const override;
     size_t getSize() const override;
 public:
     Sequence<T>* appendImmutable( const T& value ) const override;
     Sequence<T>* prependImmutable( const T& value ) const override;
-    Sequence<T>* insertAtImmutable( const T& value, const int pos ) const override;
-    Sequence<T>* removeAtImmutable( const int pos ) const override;
-    Sequence<T>* setAtImmutable( const T& value, const int pos ) const override;
-    Sequence<T>* swapImmutable( const int pos1, const int pos2 ) const override;
+    Sequence<T>* insertAtImmutable( const T& value, const size_t pos ) const override;
+    Sequence<T>* removeAtImmutable( const size_t pos ) const override;
+    Sequence<T>* setAtImmutable( const T& value, const size_t pos ) const override;
+    Sequence<T>* swapImmutable( const size_t pos1, const size_t pos2 ) const override;
     Sequence<T>* concatImmutable( const Sequence<T>& other ) const override;
     Sequence<T>* mapImmutable( const std::function<T(T)>& func ) const;
     Sequence<T>* whereImmutable( const std::function<bool(T)>& func ) const;
