+++ inc/utilInc/Sequence.hpp
@@ -16,15 +16,15 @@ public:
 public:
     virtual void append( const T& value ) = 0;
     virtual void prepend( const T& value ) = 0;
-    virtual void insertAt( const T& value, const int pos ) = 0;
-    virtual void removeAt( const int pos ) = 0;
-    virtual void setAt( const T& value, const int pos ) = 0;
-    virtual void swap( const int pos1, const int pos2 ) = 0;
-    virtual Sequence<T>* getSubSequence( const int startIndex, const int endIndex ) const = 0;
+    virtual void insertAt( const T& value, const size_t pos ) = 0;
+    virtual void removeAt( const size_t pos ) = 0;
+    virtual void setAt( const T& value, const size_t pos ) = 0;
+    virtual void swap( const size_t pos1, const size_t pos2 ) = 0;
+    virtual Sequence<T>* getSubSequence( const size_t startIndex, const size_t endIndex ) const = 0;
     virtual Sequence<T>* concat( const Sequence<T>& other ) = 0;
 public:
-    virtual T& operator[]( const int pos ) = 0;
-    virtual const T& operator[]( const int pos ) const = 0;
+    virtual T& operator[]( const size_t pos ) = 0;
+    virtual const T& operator[]( const size_t pos ) const = 0;
 public:
     virtual bool isEmpty() const = 0;
     virtual size_t getSize() const = 0;
@@ -33,10 +33,10 @@ public:
 public: // immutable functions
     virtual Sequence<T>* appendImmutable( const T& value ) const = 0;
     virtual Sequence<T>* prependImmutable( const T& value ) const = 0;
-    virtual Sequence<T>* insertAtImmutable( const T& value, const int pos ) const = 0;
-    virtual Sequence<T>* removeAtImmutable( const int pos ) const = 0;
-    virtual Sequence<T>* setAtImmutable( const T& value, const int pos ) const = 0;
-    virtual Sequence<T>* swapImmutable( const int pos1, const int pos2 ) const = 0;
+    virtual Sequence<T>* insertAtImmutable( const T& value, const size_t pos ) const = 0;
+    virtual Sequence<T>* removeAtImmutable( const size_t pos ) const = 0;
+    virtual Sequence<T>* setAtImmutable( const T& value, const size_t pos ) const = 0;
+    virtual Sequence<T>* swapImmutable( const size_t pos1, const size_t pos2 ) const = 0;
     virtual Sequence<T>* concatImmutable( const Sequence<T>& other ) const = 0;
 };
